04_ml_pipeline_dvc/requirements.txt
numpy
pandas
scikit-learn
pyyaml
requests
dvc[all]
mlflow
